+++ R/excel-stat-mutation-functions.R
@@ -29,6 +29,7 @@
 #' @examples
 #' # Libraries
 #' library(tidyquant)
+#' library(timetk)
 #' library(tidyverse)
 #' library(forcats)
 #'
@@ -49,7 +50,7 @@
 #'     # Summarization - Collapse from daily to FIRST value by month
 #'     summarise_by_time(
 #'         .date_var  = date,
-#'         .time_unit = "month",
+#'         .by        = "month",
 #'         adjusted   = FIRST(adjusted)
 #'     ) %>%
 #'