+++ R/excel-stat-summary-functions.R
@@ -39,6 +39,7 @@
 #' @examples
 #' # Libraries
 #' library(tidyquant)
+#' library(timetk)
 #' library(tidyverse)
 #' library(forcats)
 #'
@@ -59,7 +60,7 @@
 #'     # Summarization - Collapse from daily to FIRST value by month
 #'     summarise_by_time(
 #'         .date_var  = date,
-#'         .time_unit = "month",
+#'         .by        = "month",
 #'         adjusted   = FIRST(adjusted)
 #'     )
 #'