+++ R/ggplot-theme_tq.R
@@ -35,6 +35,7 @@ NULL
 
 #' @rdname theme_tq
 #' @export
+#' @importFrom ggplot2 `%+replace%`
 theme_tq <- function(base_size = 11, base_family = "") {
 
     # Tidyquant colors
@@ -44,7 +45,7 @@ theme_tq <- function(base_size = 11, base_family = "") {
     grey  <- "grey80"
 
     # Starts with theme_grey and then modify some parts
-    theme_grey(base_size = base_size, base_family = base_family) %+replace%
+    ggplot2::theme_grey(base_size = base_size, base_family = base_family) %+replace%
         ggplot2::theme(
 
             # Base Inherited Elements
@@ -59,15 +60,15 @@ theme_tq <- function(base_size = 11, base_family = "") {
 
             # Axes
             axis.line          = ggplot2::element_blank(),
-            axis.text          = ggplot2::element_text(size = rel(0.8)),
-            axis.ticks         = ggplot2::element_line(color = grey, size = rel(1/3)),
-            axis.title         = ggplot2::element_text(size = rel(1.0)),
+            axis.text          = ggplot2::element_text(size = ggplot2::rel(0.8)),
+            axis.ticks         = ggplot2::element_line(color = grey, size = ggplot2::rel(1/3)),
+            axis.title         = ggplot2::element_text(size = ggplot2::rel(1.0)),
 
             # Panel
             panel.background   = ggplot2::element_rect(fill = white, color = NA),
-            panel.border       = ggplot2::element_rect(fill = NA, size = rel(1/2), color = blue),
-            panel.grid.major   = ggplot2::element_line(color = grey, size = rel(1/3)),
-            panel.grid.minor   = ggplot2::element_line(color = grey, size = rel(1/3)),
+            panel.border       = ggplot2::element_rect(fill = NA, size = ggplot2::rel(1/2), color = blue),
+            panel.grid.major   = ggplot2::element_line(color = grey, size = ggplot2::rel(1/3)),
+            panel.grid.minor   = ggplot2::element_line(color = grey, size = ggplot2::rel(1/3)),
             panel.grid.minor.x = ggplot2::element_blank(),
             panel.spacing      = ggplot2::unit(.75, "cm"),
 
@@ -80,9 +81,9 @@ theme_tq <- function(base_size = 11, base_family = "") {
             strip.text         = ggplot2::element_text(color = white, size = ggplot2::rel(0.8), margin = ggplot2::margin(t = 5, b = 5)),
 
             # Plot
-            plot.title         = ggplot2::element_text(size = rel(1.2), hjust = 0,
+            plot.title         = ggplot2::element_text(size = ggplot2::rel(1.2), hjust = 0,
                                                        margin = ggplot2::margin(t = 0, r = 0, b = 4, l = 0, unit = "pt")),
-            plot.subtitle      = ggplot2::element_text(size = rel(0.9), hjust = 0,
+            plot.subtitle      = ggplot2::element_text(size = ggplot2::rel(0.9), hjust = 0,
                                                        margin = ggplot2::margin(t = 0, r = 0, b = 3, l = 0, unit = "pt")),
 
             # Complete theme
@@ -105,12 +106,12 @@ theme_tq_dark <- function(base_size = 11, base_family = "") {
         ggplot2::theme(
 
             # Axes
-            axis.ticks         = ggplot2::element_line(color = blue, size = rel(1/3)),
+            axis.ticks         = ggplot2::element_line(color = blue, size = ggplot2::rel(1/3)),
 
             # Panel
             panel.background   = ggplot2::element_rect(fill = grey, color = NA),
-            panel.grid.major   = ggplot2::element_line(color = white, size = rel(1/3)),
-            panel.grid.minor   = ggplot2::element_line(color = white, size = rel(1/3)),
+            panel.grid.major   = ggplot2::element_line(color = white, size = ggplot2::rel(1/3)),
+            panel.grid.minor   = ggplot2::element_line(color = white, size = ggplot2::rel(1/3)),
 
             # Complete theme
             complete = TRUE
@@ -132,12 +133,12 @@ theme_tq_green <- function(base_size = 11, base_family = "") {
         ggplot2::theme(
 
             # Axes
-            axis.ticks         = ggplot2::element_line(color = blue, size = rel(1/3)),
+            axis.ticks         = ggplot2::element_line(color = blue, size = ggplot2::rel(1/3)),
 
             # Panel
             panel.background   = ggplot2::element_rect(fill = green, color = NA),
-            panel.grid.major   = ggplot2::element_line(color = white, size = rel(1/3)),
-            panel.grid.minor   = ggplot2::element_line(color = white, size = rel(1/3)),
+            panel.grid.major   = ggplot2::element_line(color = white, size = ggplot2::rel(1/3)),
+            panel.grid.minor   = ggplot2::element_line(color = white, size = ggplot2::rel(1/3)),
 
             # Complete theme
             complete = TRUE